requirements.txt
PyYAML 5.3.1
pytest 6.1.0
selenium 3.141.0